src/requirements.txt
flask
flask-migrate
flask-sqlalchemy
gunicorn
psycopg2-binary
python-decouple
python-dotenv
